pyproject.toml
[build-system]
requires = ["setuptools>=82.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wpt-gen"
version = "0.1.0"
description = "A CLI Tool for AI-Powered Web Platform Test Generation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core LLM Integration
    "google-genai>=1.62.0",
    "openai>=2.20.0",

    # Context & Scraping
    "trafilatura>=2.0.0",

    # Prompt Templating
    "jinja2>=3.1.6",

    # CLI Infrastructure
    "typer>=0.23.0",
    "rich>=14.3.2",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.0",
    "mypy>=1.19.1",
]

[project.scripts]
wpt-gen = "wptgen.main:app"

[tool.setuptools.packages.find]
include = ["wptgen*"]

[tool.setuptools.package-data]
wptgen = ["templates/*.jinja", "templates/*.xml", "templates/*.txt"]

[tool.ruff]
# Target Python 3.10+
target-version = "py310"
line-length = 100
indent-width = 2

[tool.ruff.lint]
# Enable a wide set of rules for "strict" linting
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "PT",   # flake8-pytest-style
]
ignore = [
    "E501", # Line too long (already handled by line-length and formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["wptgen"]
